refactor(product-components): split AssetLogo and make coingeckoId optional
