+++ packages/suite/src/components/earn/dashboard/common/EarnAccountCell.tsx
@@ -2,9 +2,7 @@ import { type TokenDto } from '@suite-common/earn-api';
 import { type NetworkSymbol } from '@suite-common/wallet-config';
 import { type Account } from '@suite-common/wallet-types';
 import { Column, Row } from '@trezor/components';
-import { CoinLogo } from '@trezor/product-components';
-
-import { VaultTokenLogo } from 'src/components/earn/common/VaultTokenLogo';
+import { AssetLogo } from '@trezor/product-components';
 
 import { EarnAccountCellDetails } from './EarnAccountCellDetails';
 import { type EarnTokenBalance } from './types';
@@ -33,16 +31,14 @@ export const EarnAccountCell = ({
     return (
         <Row gap={16} cursor="inherit">
             <Column alignItems="center">
-                {iconToken ? (
-                    <VaultTokenLogo
-                        token={iconToken}
-                        networkSymbol={networkSymbol}
-                        size={32}
-                        showNetworkIcon={showAssetNetworkIcon}
-                    />
-                ) : (
-                    <CoinLogo size={32} symbol={networkSymbol} type="tokenWithNetwork" />
-                )}
+                <AssetLogo
+                    coingeckoId={iconToken?.coinGeckoId ?? undefined}
+                    placeholder={iconToken?.symbol || iconToken?.name || ''}
+                    symbol={networkSymbol}
+                    contractAddress={iconToken?.address ?? null}
+                    showNetworkIcon={showAssetNetworkIcon}
+                    size={32}
+                />
             </Column>
 
             <Column flex="1" overflow="hidden" gap={2}>
