+++ packages/suite/src/components/earn/dashboard/yield/EarnYieldApyBreakdown.tsx
@@ -1,14 +1,27 @@
-import { Translation } from '@suite/intl';
-import { type RewardDto } from '@suite-common/earn-api';
+import { Translation, type TranslationKey } from '@suite/intl';
+import { type RewardDto, type RewardDtoYieldSource } from '@suite-common/earn-api';
 import { type NetworkSymbol } from '@suite-common/wallet-config';
 import { Column, Icon, Row, Text } from '@trezor/components';
+import { AssetLogo } from '@trezor/product-components';
 
-import { VaultTokenLogo } from 'src/components/earn/common/VaultTokenLogo';
 import { getApyPercent } from 'src/components/earn/utils/earnApyUtils';
 
 type EarnYieldApyBreakdownProps = {
     rewards: RewardDto[];
     networkSymbol: NetworkSymbol;
+};
+
+const getYieldSourceTranslationId = (yieldSource: RewardDtoYieldSource): TranslationKey | null => {
+    switch (yieldSource) {
+        case 'vault':
+        case 'lending_interest':
+            return 'TR_EARN_YIELD_APY_SOURCE_LENDING_INTEREST';
+        case 'protocol_incentive':
+        case 'points':
+            return 'TR_EARN_YIELD_APY_SOURCE_PROTOCOL_INCENTIVE';
+        default:
+            return null;
+    }
 };
 
 const sortRewards = (rewards: RewardDto[]) =>
@@ -23,44 +36,30 @@ export const EarnYieldApyBreakdown = ({ rewards, networkSymbol }: EarnYieldApyBr
     <Column gap={16} padding={{ vertical: 10, horizontal: 8 }}>
         {sortRewards(rewards).map((reward, index) => {
             const ratePercent = getApyPercent(reward.rate);
+            const translationId = getYieldSourceTranslationId(reward.yieldSource);
+            const description = translationId ? (
+                <Translation id={translationId} />
+            ) : (
+                reward.description
+            );
 
             return (
                 <Row key={index} gap={8} alignItems="center">
-                    <VaultTokenLogo
-                        token={reward.token}
-                        networkSymbol={networkSymbol}
-                        size={20}
+                    <AssetLogo
+                        coingeckoId={reward.token.coinGeckoId}
+                        placeholder={reward.token.symbol || reward.token.name || 'token'}
+                        symbol={networkSymbol}
+                        contractAddress={reward.token.address}
                         showNetworkIcon
+                        size={20}
                     />
                     <Column flex="1">
                         <Text typographyStyle="body-sm">{reward.token.symbol}</Text>
-                        {reward.description && (
+                        {description && (
                             <Text typographyStyle="body-sm" intent="neutral" priority="secondary">
-                                {reward.description}
+                                {description}
                             </Text>
                         )}
-                        {!reward.description &&
-                            (reward.yieldSource === 'protocol_incentive' ||
-                                reward.yieldSource === 'points') && (
-                                <Text
-                                    typographyStyle="body-sm"
-                                    intent="neutral"
-                                    priority="secondary"
-                                >
-                                    <Translation id="TR_EARN_YIELD_APY_SOURCE_PROTOCOL_INCENTIVE" />
-                                </Text>
-                            )}
-                        {!reward.description &&
-                            reward.yieldSource !== 'protocol_incentive' &&
-                            reward.yieldSource !== 'points' && (
-                                <Text
-                                    typographyStyle="body-sm"
-                                    intent="neutral"
-                                    priority="secondary"
-                                >
-                                    <Translation id="TR_EARN_YIELD_APY_SOURCE_LENDING_INTEREST" />
-                                </Text>
-                            )}
                     </Column>
                     {ratePercent !== null && ratePercent > 0 && (
                         <Text typographyStyle="body-sm" intent="brand">
